Move ArcInterval to dht_arc module, DhtLocation to dht_location module
